complete exercise 4.21

+++ controllers/blogs.js
@@ -29,8 +29,25 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async (request, response) => {
     logger.info('deleting blog entry with id', request.params.id);
-    await Blog.findByIdAndDelete(request.params.id);
-    return response.status(204).end();
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if(!decodedToken.id) {
+        return response.status(401).json({ error:'invalid token' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    if(blog){
+        if(blog.user.toString() === decodedToken.id){
+            await Blog.findByIdAndDelete(request.params.id);
+            return response.status(204).end();
+        }
+        else{
+            return response.status(401).json({ error:'you are unauthorized to delete this blog entry' });
+        }
+    }
+    else {
+        return response.status(204).end();
+    }
 })
 
 blogRouter.put('/:id', async (request, response) => {
